eth_perp_ofi_sim_realtime/requirements.txt
pandas==2.2.2
numpy==1.26.4
pyyaml==6.0.2
websockets==12.0
